#ID-1262 Updated work with pages

+++ gateway/modules/admin/controllers/traits/settings/PagesTrait.php
@@ -178,6 +178,10 @@ trait PagesTrait {
          */
         $pageModel = $this->_findModel($id, Pages::class);
 
+        if (!$pageModel->can(Pages::CAN_DELETE)) {
+            return [false];
+        }
+
         $pageModel->delete();
 
         UiHelper::message(Yii::t('admin', 'settings.pages_message_deleted'));
